# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: MyWeb E-commerce API
  version: "1.0.0"
  description: |
    REST API for MyWeb ecommerce (clothes, dresses, shoes).
    
    **Auth**: JWT Bearer. Public endpoints: `/api/v1/health`, `/api/v1/auth/**`, `/api/v1/categories/**`, `/api/v1/products/**`.
    All others require JWT. Admin endpoints additionally require `ROLE_ADMIN`.
servers:
  - url: http://localhost:8080
    description: Local dev
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      in: query
      name: page
      schema:
        type: integer
        minimum: 0
        default: 0
      description: 0-based page index
    SizeParam:
      in: query
      name: size
      schema:
        type: integer
        minimum: 1
        default: 20
      description: Page size
    ProductSortParam:
      in: query
      name: sortBy
      schema:
        type: string
        enum: [newest, oldest, priceAsc, priceDesc]
        default: newest
      description: Sort option
    ProductKeywordParam:
      in: query
      name: keyword
      schema:
        type: string
      description: Full-text keyword for products
    ProductCategoryParam:
      in: query
      name: category
      schema:
        type: string
      description: Category slug filter
    CommonStatusParam:
      in: query
      name: status
      schema:
        type: string
      description: Optional status filter
  responses:
    Unauthorized:
      description: Missing/invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Not allowed (e.g., missing ADMIN role)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string

    # ---- Auth / User ----
...
    AdminProductRequest:
      type: object
      required: [name, slug, categoryId, basePrice]
      properties:
        name:
          type: string
        slug:
          type: string
        shortDesc:
          type: string
        longDesc:
          type: string
        categoryId:
          type: integer
          format: int64
        basePrice:
          type: number
          format: double
        stockQuantity:
          type: integer
        thumbnailUrl:
          type: string
        status:
          type: string
          enum: [ACTIVE, INACTIVE, DRAFT]
        imageUrls:
          type: array
          items:
            type: string
          description: List of additional product image URLs
    AdminProductResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        slug:
          type: string
        shortDesc:
          type: string
        longDesc:
          type: string
        categoryId:
          type: integer
          format: int64
        categoryName:
          type: string
        categorySlug:
          type: string
        basePrice:
          type: number
          format: double
        stockQuantity:
          type: integer
        thumbnailUrl:
          type: string
        status:
          type: string
          enum: [ACTIVE, INACTIVE, DRAFT]
        imageUrls:
          type: array
          items:
            type: string
          description: List of additional product image URLs
    ProductImageResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        imageUrl:
          type: string
        thumbnail:
          type: boolean
        sortOrder:
          type: integer
    ProductVariantResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        sku:
          type: string
        color:
          type: string
        size:
          type: string
        price:
          type: number
          format: double
          nullable: true
        stock:
          type: integer
        status:
          type: string
...
    ProductListItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        slug:
          type: string
        thumbnailUrl:
          type: string
        price:
          type: number
          format: double
        categorySlug:
          type: string
        categoryName:
          type: string
        createdAt:
          type: string
          format: date-time
    ProductSuggestionResponse:
      type: object
      description: Simplified product data for search autocomplete/suggestions
      properties:
        id:
          type: integer
          format: int64
          description: Product ID
        name:
          type: string
          description: Product name
        slug:
          type: string
          description: URL-friendly slug for navigation
        thumbnailUrl:
          type: string
          description: Product thumbnail image URL
        price:
          type: number
          format: double
          description: Product base price
        categoryName:
          type: string
          description: Category name for context
...
    ProductDetailResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        slug:
          type: string
        shortDesc:
          type: string
        longDesc:
          type: string
        categorySlug:
          type: string
        categoryName:
          type: string
        basePrice:
          type: number
          format: double
        thumbnailUrl:
          type: string
        status:
          type: string
          enum: [ACTIVE, INACTIVE, DRAFT]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        images:
          type: array
          items:
            $ref: '#/components/schemas/ProductImageResponse'
        variants:
          type: array
          items:
            $ref: '#/components/schemas/ProductVariantResponse'

    # ---- Cart ----
...
    OrderItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        productId:
          type: integer
          format: int64
        variantId:
          type: integer
          format: int64
          nullable: true
        productName:
          type: string
        productSlug:
          type: string
        thumbnailUrl:
          type: string
        color:
          type: string
        size:
          type: string
        unitPrice:
          type: number
          format: double
        quantity:
          type: integer
        totalPrice:
          type: number
          format: double
...
    OrderDetailResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        orderNumber: { type: string }
        createdAt: { type: string, format: date-time }
        status: { type: string }
        timeline:
          type: array
          items:
            type: object
            properties:
              status: { type: string }
              timestamp: { type: string, format: date-time }
              completed: { type: boolean }
        items:
          type: array
          items:
            type: object
            properties:
              id: { type: integer, format: int64 }
              productName: { type: string }
              productImageUrl: { type: string }
              variantInfo: { type: string }
              quantity: { type: integer }
              price: { type: number, format: double }
        subtotal: { type: number, format: double }
        shippingFee: { type: number, format: double }
        totalAmount: { type: number, format: double }
        shippingAddress:
          type: object
          properties:
            fullName: { type: string }
            addressLine: { type: string }
            phone: { type: string }
        paymentMethod:
          type: object
          properties:
            type: { type: string }
            last4: { type: string }
    OrderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        status:
          type: string
          enum: [PENDING, CONFIRMED, PACKED, SHIPPED, DELIVERED, CANCELED, RETURNED, REFUNDED]
        paymentStatus:
          type: string
          enum: [UNPAID, PAID, FAILED, REFUNDED]
        paymentMethod:
          type: string
          enum: [COD, VNPAY, MOMO]
        totalAmount:
          type: number
          format: double
        shippingFee:
          type: number
          format: double
        note:
          type: string
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
    AdminOrderSummaryResponse:
      type: object
      description: Summary used in /api/v1/admin/orders (dto.order)
      properties:
        id:
          type: integer
          format: int64
        userEmail:
          type: string
        shippingFullName:
          type: string
        totalAmount:
          type: number
          format: double
        shippingFee:
          type: number
          format: double
        status:
          type: string
          enum: [PENDING, CONFIRMED, PACKED, SHIPPED, DELIVERED, CANCELED, RETURNED, REFUNDED]
        paymentMethod:
          type: string
          enum: [COD, VNPAY, MOMO]
        paymentStatus:
          type: string
          enum: [UNPAID, PAID, FAILED, REFUNDED]
        createdAt:
          type: string
          format: date-time
    AdminOrderItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        productId:
          type: integer
          format: int64
        productName:
          type: string
        unitPrice:
          type: number
          format: double
        quantity:
          type: integer
        lineTotal:
          type: number
          format: double
    AdminOrderDetailResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        status:
          type: string
          enum: [PENDING, PAID, SHIPPED, COMPLETED, CANCELED]
        paymentMethod:
          type: string
          enum: [COD, VNPAY, MOMO]
        paymentStatus:
          type: string
          enum: [UNPAID, PAID, FAILED, REFUNDED]
        totalAmount:
          type: number
          format: double
        shippingFee:
          type: number
          format: double
        note:
          type: string
        userId:
          type: integer
          format: int64
        userEmail:
          type: string
        shippingFullName:
          type: string
        shippingPhone:
          type: string
        shippingAddressLine1:
          type: string
        shippingWard:
          type: string
        shippingDistrict:
          type: string
        shippingProvince:
          type: string
        shippingCountry:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminOrderItemResponse'
    UpdateOrderStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [PENDING, CONFIRMED, PACKED, SHIPPED, DELIVERED, CANCELED, RETURNED, REFUNDED]

    # ---- Payment ----
...
    Page_ProductListItemResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProductListItemResponse'
        number:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
...
    Page_AdminOrderSummaryResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AdminOrderSummaryResponse'
        number:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
...
paths:
...
  /api/v1/users/me/orders/{orderId}:
    get:
      tags:
        - "CLIENT - Orders"
      summary: Get my order detail
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/products:
    get:
      tags:
        - "CLIENT - Products & Catalog"
      security: []
      summary: List products (public)
      parameters:
        - $ref: '#/components/parameters/ProductCategoryParam'
        - $ref: '#/components/parameters/ProductKeywordParam'
        - in: query
          name: minPrice
          schema:
            type: number
            format: double
          description: Minimum price filter
        - in: query
          name: maxPrice
          schema:
            type: number
            format: double
          description: Maximum price filter
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/ProductSortParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_ProductListItemResponse'
  /api/v1/products/slug/{slug}:
    get:
      tags:
        - "CLIENT - Products & Catalog"
      security: []
      summary: Get product by slug (public)
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetailResponse'
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/products/{id}:
    get:
      tags:
        - "CLIENT - Products & Catalog"
      security: []
      summary: Get product by id (public)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetailResponse'
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/products/{id}/related:
    get:
      tags:
        - "CLIENT - Products & Catalog"
      security: []
      summary: Get related products (same category)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_ProductListItemResponse'
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/products/suggestions:
    get:
      tags:
        - "CLIENT - Products & Catalog"
      security: []
      summary: Get search suggestions for autocomplete (public)
      description: |
        Returns product suggestions for search autocomplete.
        Can be filtered by category for category-specific suggestions.
      parameters:
        - in: query
          name: keyword
          schema:
            type: string
          description: Search keyword (minimum 2 characters recommended)
          example: "head"
        - in: query
          name: category
          schema:
            type: string
          description: Category slug to filter suggestions (fashion, beauty, electronics, home-living)
          example: "electronics"
        - in: query
          name: limit
          schema:
            type: integer
            default: 5
            minimum: 1
            maximum: 20
          description: Maximum number of suggestions to return
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductSuggestionResponse'
              example:
                - id: 123
                  name: "Premium Wireless Headphones 5"
                  slug: "premium-wireless-headphones-5"
                  thumbnailUrl: "https://images.unsplash.com/photo-1505740420928-5e560c06d30e"
                  price: 289.99
                  categoryName: "Electronics"
                - id: 124
                  name: "Ultra Wireless Headphones 12"
                  slug: "ultra-wireless-headphones-12"
                  thumbnailUrl: "https://images.unsplash.com/photo-1505740420928-5e560c06d30e"
                  price: 299.99
                  categoryName: "Electronics"
...
  /api/v1/admin/products/{id}:
    get:
      tags:
        - "ADMIN - Product Management"
      summary: Get product detail (ADMIN)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminProductResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags:
        - "ADMIN - Product Management"
      summary: Update product (ADMIN)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminProductRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminProductResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags:
        - "ADMIN - Product Management"
      summary: Delete product (ADMIN)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---- Admin: Orders ----
  /api/v1/admin/orders:
    get:
      tags:
        - "ADMIN - Order Management"
      summary: List orders (ADMIN)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/CommonStatusParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_AdminOrderSummaryResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/admin/orders/{orderId}:
    get:
      tags:
        - "ADMIN - Order Management"
      summary: Order detail (ADMIN)
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOrderDetailResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/admin/orders/{orderId}/status:
    patch:
      tags:
        - "ADMIN - Order Management"
      summary: Update order status (ADMIN)
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOrderDetailResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---- Admin: Users ----
...
  /api/v1/admin/orders/{id}/return/approve:
    put:
      tags:
        - "ADMIN - Order Management"
      summary: Approve return request (ADMIN)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Return approved, stock replenished
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/admin/orders/{id}/status:
    put:
      tags:
        - "ADMIN - Order Management"
      summary: Update order status (ADMIN)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateOrderStatusRequest' }
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderResponse' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---- Admin Reviews ----
...